feat: memset int to longlong

+++ get_next_line_bonus.h
@@ -6,7 +6,7 @@
 /*   By: hyunwkim <hyunwkim@42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/06/22 10:09:47 by hyunwkim          #+#    #+#             */
-/*   Updated: 2021/06/26 18:32:12 by hyunwkim         ###   ########.fr       */
+/*   Updated: 2021/06/28 12:23:31 by hyunwkim         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ char			*ft_strdup(char *s1);
 int				ft_strlen(char *s);
 int				ft_strlcpy(char *dst, char *src, int dstsize);
 char			*ft_strjoin(char *s1, char *s2, int read_size);
-void			*ft_memset(char *b, int c, int len);
+void			*ft_memset(char *b, int c, long long len);
 int				is_newline(char *bak);
 int				alloc_newline(char **line, char **bak, int nl_idx, int read);
 int				get_next_line(int fd, char **line);
